Move FieldSortSelect to views components widget directory
