+++ graylog2-web-interface/src/views/components/widgets/EditMessageList.jsx
@@ -7,7 +7,7 @@ import * as Immutable from 'immutable';
 
 import SortableSelect from 'views/components/aggregationbuilder/SortableSelect';
 import CustomPropTypes from 'views/components/CustomPropTypes';
-import FieldSortSelect from 'views/components/FieldSortSelect';
+import FieldSortSelect from 'views/components/widgets/FieldSortSelect';
 import SortDirectionSelect from 'views/components/aggregationbuilder/SortDirectionSelect';
 
 import AggregationWidgetConfig from 'views/logic/aggregationbuilder/AggregationWidgetConfig';
