+++ graylog2-web-interface/src/views/components/widgets/FieldSortSelect.jsx
